split StateTests wip
